requirements.txt
# Backend Framework
quart>=0.19.0
quart-cors>=0.7.0
hypercorn>=0.16.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP Client
httpx>=0.27.0

# Scheduler
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0

# Security
cryptography>=42.0.0

# Date handling
pytz>=2024.1
